refactor: 메뉴그룹 루트 애그리거트를 패키징하라
