+++ src/main/java/kitchenpos/menus/menugroup/domain/JpaMenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.domain;
+package kitchenpos.menus.menugroup.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
