+++ src/main/java/kitchenpos/menus/menugroup/domain/MenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.domain;
+package kitchenpos.menus.menugroup.domain;
 
 import java.util.List;
 import java.util.Optional;
